refactor: create the install-tl module
* also add some missing tests, and
* make sure to add tlcontrib even when using a cache
