+++ .eslintrc
@@ -97,6 +97,7 @@ rules:
 overrides:
   - files:
       - ./__tests__/context.test.ts
+      - ./__tests__/install-tl.test.ts
       - ./__tests__/setup.test.ts
       - ./__tests__/texlive.test.ts
     extends:
